centralized the validate rules
